Sun damage >>>> Spectral traps
instead of using a "spectral trap" that was a box which causes damage to simulate the sun, now directly a ray ticks every .5 sec from the directional light to the player. Besides that, this new way adapts to the irregular forms of the environment, so we are not limited to use squared lightning/shadows.